Add input generator for load testing
A Python script has been added with configurable parameters for the generation of sample data to use in performance testing. This generator creates JSON-formatted facts and rules files, as well as a Rule Engine configuration file in YAML format. Additionally, support for a Path object in addition to string file paths has been added in RuleEngineConfigProducer's to_yaml function.